Move the text representation tests into their own file
They assert appearance rather than conduct, so keeping them next to the
functional tests made every printout change look like a behavior change. The
four printouts and the two that prove an unset field renders no line now live in
test_text_representations.py, and the four variants that only checked a different
value interpolate are gone, since the kept printouts cover that already.
