+++ tests/test_minimal/test_tools/test_backend_and_dataset_configuration/test_models/test_zarr_dataset_io_configuration_model.py
@@ -1,8 +1,5 @@
 """Unit tests for the ZarrDatasetIOConfiguration Pydantic model."""
 
-from io import StringIO
-from unittest.mock import patch
-
 import pytest
 from numcodecs import GZip
 
@@ -11,74 +8,6 @@ from neuroconv.tools.nwb_helpers import (
     ZarrDatasetIOConfiguration,
 )
 from neuroconv.tools.testing import mock_ZarrDatasetIOConfiguration
-
-
-def test_zarr_dataset_io_configuration_print():
-    """Test the printout display of a ZarrDatasetIOConfiguration model looks nice."""
-    zarr_dataset_configuration = mock_ZarrDatasetIOConfiguration()
-
-    with patch("sys.stdout", new=StringIO()) as out:
-        print(zarr_dataset_configuration)
-
-    expected_print = """
-acquisition/TestElectricalSeries/data
--------------------------------------
-  dtype : int16
-  full shape of source array : (1800000, 384)
-  full size of source array : 1.38 GB
-
-  buffer shape : (1250000, 384)
-  expected RAM usage : 960.00 MB
-
-  chunk shape : (78125, 64)
-  disk space usage per chunk : 10.00 MB
-
-  compression method : gzip
-
-"""
-    assert out.getvalue() == expected_print
-
-
-def test_zarr_dataset_configuration_print_with_compression_options():
-    """The compression options line renders only when options are set."""
-    zarr_dataset_configuration = mock_ZarrDatasetIOConfiguration(compression_options=dict(level=5))
-
-    printout = str(zarr_dataset_configuration)
-
-    assert "  compression method : gzip" in printout
-    assert "  compression options : {'level': 5}" in printout
-
-
-def test_zarr_dataset_configuration_print_with_compression_disabled():
-    """Neither compression line renders when compression is disabled."""
-    zarr_dataset_configuration = mock_ZarrDatasetIOConfiguration(compression_method=None)
-
-    printout = str(zarr_dataset_configuration)
-
-    assert "compression method" not in printout
-    assert "compression options" not in printout
-
-
-def test_zarr_dataset_configuration_print_with_filter_methods():
-    """The filter methods line renders only when filters are set."""
-    zarr_dataset_configuration = mock_ZarrDatasetIOConfiguration(filter_methods=["delta", "blosc"])
-
-    printout = str(zarr_dataset_configuration)
-
-    assert "  filter methods : ['delta', 'blosc']" in printout
-    assert "filter options" not in printout
-
-
-def test_zarr_dataset_configuration_print_with_filter_options():
-    """The filter options line renders alongside the filter methods line."""
-    zarr_dataset_configuration = mock_ZarrDatasetIOConfiguration(
-        filter_methods=["blosc"], filter_options=[dict(clevel=6)]
-    )
-
-    printout = str(zarr_dataset_configuration)
-
-    assert "  filter methods : ['blosc']" in printout
-    assert "  filter options : [{'clevel': 6}]" in printout
 
 
 def test_validator_filter_options_has_methods():
